大幅修正
Grid.cpp修正

+++ capturability/include/base.h
@@ -11,8 +11,8 @@ typedef Eigen::Vector2f vec2_t;
 typedef Eigen::Vector3f vec3_t;
 typedef std::vector<vec2_t> arr2_t;
 
-float  dot(vec2_t v1, vec2_t v2);
-float  cross(vec2_t v1, vec2_t v2);
+double  dot(vec2_t v1, vec2_t v2);
+double  cross(vec2_t v1, vec2_t v2);
 vec2_t normal(vec2_t v);
 
 int round(double val);
