+++ capturability/include/swing_foot.h
@@ -15,16 +15,16 @@ public:
 
   void   set(vec2_t foot, vec2_t foot_des);
   void   set(vec3_t foot, vec3_t foot_des);
-  float  getTime();
-  vec3_t getTraj(float dt);
+  double  getTime();
+  vec3_t getTraj(double dt);
 
 private:
   Cycloid cycloid;
 
   vec3_t foot, foot_des;
-  float  step_time_min, step_time;
-  float  step_height;
-  float  foot_vel;
+  double  step_time_min, step_time;
+  double  step_height;
+  double  foot_vel;
 };
 
 } // namespace Capt
